requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.29
psycopg2-binary==2.9.9
requests==2.31.0
gunicorn==21.2.0
pytesseract==0.3.10
Pillow==10.2.0
